Remove repository interfaces and provider
