+++ app/Repositories/AnnouncementRepository.php
@@ -5,12 +5,25 @@ namespace App\Repositories;
 use App\Models\Announcement;
 use Illuminate\Database\Eloquent\Collection;
 
-interface AnnouncementRepository extends Repository
+class AnnouncementRepository extends Repository
 {
+    /**
+     * EloquentRepository constructor.
+     *
+     * @param Announcement $model
+     */
+    public function __construct(Announcement $model)
+    {
+        $this->model = $model;
+    }
+
     /**
      * Return all published announcements.
      *
      * @return Collection|Announcement[]
      */
-    public function getPublished();
+    public function getPublished()
+    {
+        return $this->model->published()->get();
+    }
 }
